Refactor constants module
