+++ amset/constants.py
@@ -16,6 +16,7 @@ e = physical_constants["elementary charge"][0]
 k_B = physical_constants["Boltzmann constant in eV/K"][0]
 default_small_e = 1.0  # eV/cm the value of this parameter does not matter
 over_sqrt_pi = 1 / np.sqrt(np.pi)
+sqrt2 = np.sqrt(2)
 small_val = 1e-32  # e.g. used for an initial non-zero val
 output_width = 69
 spin_name = {Spin.up: "spin-up", Spin.down: "spin-down"}
